Cone & Cillinder

+++ Raytracing/Raytracing/CommonDefines.h
@@ -15,7 +15,7 @@
 #define DONTDISCARD	[[nodiscard]]
 #define CLASS_EXCEPTION(name) struct name ## Exception : public std::exception { name ## Exception(const char* what) : std::exception(what) {}};
 
-constexpr float PI = 3.14159265358979323846f;
+constexpr double PI = 3.14159265358979323846f;
 
 namespace std {
 	template<typename T>
