+++ Raytracing/Raytracing/Graphics/Primitives/Lighting/PointLight.h
@@ -22,10 +22,10 @@ namespace Graphics {
 #pragma endregion
 
 #pragma region //Methods
-				DONTDISCARD bool ComputeLighting(const glm::vec3& inpoint, const glm::vec3& innormal,
+				DONTDISCARD bool ComputeLighting(const glm::dvec3& inpoint, const glm::dvec3& innormal,
 					const std::vector<std::shared_ptr<Composition::Object>>& objlist,
 					const std::shared_ptr<Composition::Object>& obj,
-					glm::vec3& color, float& intensity) noexcept override;
+					glm::dvec3& color, double& intensity) noexcept override;
 #pragma endregion
 			};
 		}
