+++ Raytracing/Raytracing/Trace/Ray.cpp
@@ -22,7 +22,7 @@ namespace Trace {
 	*
 	*   Constructs a Ray with the given origin and endpoint
 	*/ // ---------------------------------------------------------------------
-	Ray::Ray(const glm::vec3& origin, const glm::vec3& endpoint) noexcept
+	Ray::Ray(const glm::dvec3& origin, const glm::dvec3& endpoint) noexcept
 		: mOrigin(origin), mEndPoint(endpoint) {
 	}
 }
